# feature(progress): adds resident set size & v8 managed memory to the performance-log