## Description

- adds _resident set size_ (rss) and the amount of C++ objects bound to javascript objects managed by v8 ('external') to the `performance-log` output.

## Motivation and Context

- provides more (useful) information to analyze what dependency-cruiser is doing under the hood.

## How Has This Been Tested?

- [x] green ci
- [x] updated unit test

## Screenshots

### after
```
$ node ./bin/dependency-cruise.js src bin test configs types tools --config --ignore-known
  elapsed       rss heapTotal  heapUsed  external after step...
    751ms     127Mb      84Mb      61Mb       2Mb start of node process
     11ms     128Mb      84Mb      62Mb       2Mb parsing options
     93ms     130Mb      85Mb      63Mb       2Mb cache: checking freshness
...
```

### before
```
$ node ./bin/dependency-cruise.js src bin test configs types tools --config --ignore-known
  elapsed heapTotal  heapUsed after step...
    751ms      84Mb      61Mb start of node process
     11ms      84Mb      62Mb parsing options
     93ms      85Mb      64Mb cache: checking freshness
...
```

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).